requirements.txt
pyspark>=3.3.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
streamlit>=1.20.0
scikit-learn>=1.2.0
kafka-python>=2.0.2
SQLAlchemy>=2.0.0
